adjusted setup.py to upload to PyPI, added python package README
