+++ python/.gitignore
@@ -1,1 +1,3 @@
+dist
 build
+neuroglancer.egg-info
