+++ python/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+description-file = README.md
